refactor : 코드 최적화, 모듈화
