+++ src/lib/carousel/Carousel.tsx
@@ -1,9 +1,9 @@
-import React, { useCallback, useRef, useState } from "react";
-import { MESSAGE } from "./constant";
+import React, { useCallback, useMemo, useRef, useState } from "react";
 import styled, { css, keyframes } from "styled-components";
 import { HiOutlineChevronLeft, HiOutlineChevronRight } from "react-icons/hi";
-import { Direction } from "./type";
-import { useCalendarDispatch, useCalendarState } from "./CalendarProvider";
+import { useCalendarDispatch } from "../calendar/CalendarProvider";
+import { MESSAGE } from "../utils/constant";
+import { Direction } from "../utils/types";
 
 type CarouselProps = {
   children: React.ReactNode;
@@ -15,13 +15,15 @@ export default function Carousel({
   countOfItemToShow,
 }: CarouselProps) {
   const [x, setX] = useState(-1);
-  const items = _createItems(children);
+  const items = useMemo(() => createItems(children), [children]);
+
   const dispatch = useCalendarDispatch();
   const moving = useRef<boolean>(false);
 
-  const onClickMove = (direction: Direction) => {
+  const onClickMove = useCallback((direction: Direction) => {
     if (moving.current) return;
     moving.current = true;
+
     if (direction === -1) {
       setX(-2);
       dispatch({ type: "MOVE_LEFT" });
@@ -29,7 +31,7 @@ export default function Carousel({
       setX(0);
       dispatch({ type: "MOVE_RIGHT" });
     }
-  };
+  }, []);
 
   const onAnimationEnd = useCallback(() => {
     moving.current = false;
@@ -48,7 +50,7 @@ export default function Carousel({
   return (
     <CarouselWarpper>
       <CarouselContainer>
-        <Slider x={x} count={countOfItemToShow} onAnimationEnd={onAnimationEnd}>
+        <Slider {...{ x, onAnimationEnd, count: countOfItemToShow }}>
           {items.map((el) => (
             <Item count={countOfItemToShow}>{el}</Item>
           ))}
@@ -74,7 +76,7 @@ Carousel.defaultProps = {
   countOfItemToShow: 2,
 };
 
-function _createItems(children: React.ReactNode) {
+function createItems(children: React.ReactNode) {
   if (!Array.isArray(children)) {
     return [children];
   }
@@ -111,16 +113,6 @@ type ControllerProps = {
   onClickHandler: (dir: Direction) => void;
 };
 
-const CarouselWarpper = styled.div`
-  width: 100%;
-  display: flex;
-`;
-
-const CarouselContainer = styled.div`
-  width: 100%;
-  /* overflow: hidden; */
-`;
-
 const moveAni = (x: number, count: number) => keyframes`
   from {
     transform: translateX(${`calc((100% / ${count}) * ${x})`});
@@ -128,6 +120,16 @@ const moveAni = (x: number, count: number) => keyframes`
   to{
     transform: translateX(${`calc((100% / ${count}) * -1)`});
   }
+`;
+
+const CarouselWarpper = styled.div`
+  width: 100%;
+  display: flex;
+`;
+
+const CarouselContainer = styled.div`
+  width: 100%;
+  overflow: hidden;
 `;
 
 const Item = styled.div<{ count: number }>`
@@ -138,7 +140,6 @@ const Item = styled.div<{ count: number }>`
 const ControllerWraaper = styled.div<{ direction: number }>`
   position: absolute;
   top: 0;
-
   ${({ direction }) =>
     direction === -1
       ? css`
@@ -159,6 +160,7 @@ const Button = styled.button`
   height: 2rem;
 
   &:hover {
-    background: tan;
+    transition: background 1s;
+    background: silver;
   }
 `;
