--- src/lib/constant.ts
@@ -1,16 +1,0 @@
-export const MESSAGE = {
-  ERROR: {
-    NO_DIRECTION:
-      "You have to send direction prop(left or right) to <Controller>",
-    BOTH_DIRECTION:
-      "<Controller> can receive only one direction prop(left or right)",
-    BOTH_TYPE: "",
-    NO_TYPE: "",
-    INVAILD_ACTION: "",
-    INVAILD_PROVIDER: "Cannot find Provider!",
-    INVAILD_RANGE_COUNT_OF_MONTH:
-      "RangeError : countOfMonth is out of range(1~12)",
-  },
-};
-
-export const DAYS = ["SON", "MON", "TUE", "WED", "THU", "FRI", "SAT"];
