+++ src/lib/utils/types.tsx
@@ -4,17 +4,25 @@ type Date = {
   day: number;
 };
 
-export type Calendar = Date & {
+export type Calendar = {
+  year: number;
+  month: number;
+  day: number;
   firstDay: number;
+};
+
+export type CurrentDay = {
+  year: number;
+  month: number;
 };
 
 export type _OnClickResult = Date & {
   week: string;
-  type: Type;
+  nextClickTarget: ClickTargetType;
 };
 
 export type Direction = 1 | -1;
-export type Type = "start" | "end";
+export type ClickTargetType = "start" | "end";
 export type OnClickDay = {
   onClickDay?: (result: _OnClickResult) => void;
 };
